+++ src/length.rs
@@ -34,62 +34,3 @@ impl Unit for LengthUnit {
         }
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::LengthUnit;
-    use crate::value::Value;
-
-    #[test]
-    fn test_factories_() {
-        let km = Value::new(LengthUnit::Kilometer, 3.0);
-        let m  = Value::new(LengthUnit::Meter, 2.0);
-        let cm = Value::new(LengthUnit::Centimeter, 50.0);
-        let mm = Value::new(LengthUnit::Millimeter, 20.0);
-
-        assert_eq!(3.0, km.value());
-        assert_eq!(3000.0, km.base_value());
-        assert_eq!(2.0, m.value());
-        assert_eq!(2.0, m.base_value());
-        assert_eq!(50.0, cm.value());
-        assert_eq!(0.5, cm.base_value());
-        assert_eq!(20.0, mm.value());
-        assert_eq!(0.02, mm.base_value());
-    }
-
-    #[test]
-    fn test_calculations() {
-        let mut m  = Value::new(LengthUnit::Meter, 2.0);
-        let cm     = Value::new(LengthUnit::Centimeter, 50.0);
-
-        let a = m + cm;
-        assert_eq!(2.5, a.value());
-        let b = m - cm;
-        assert_eq!(1.5, b.value());
-        m += cm;
-        assert_eq!(2.5, m.value());
-        m -= cm;
-        assert_eq!(2.0, m.value());
-    }
-
-    #[test]
-    fn test_strings() {
-        let km = Value::new(LengthUnit::Kilometer, 3.0);
-        let m  = Value::new(LengthUnit::Meter, 2.5);
-        let cm = Value::new(LengthUnit::Centimeter, 50.0);
-        let mm = Value::new(LengthUnit::Millimeter, 20.2);
-        assert_eq!("3km", km.to_string());
-        assert_eq!("2.5m", m.to_string());
-        assert_eq!("50cm", cm.to_string());
-        assert_eq!("20.2mm", mm.to_string());
-    }
-
-    #[test]
-    fn test_conversion() {
-        let km = Value::new(LengthUnit::Kilometer, 1.5);
-        let m = km.convert(LengthUnit::Meter);
-        let cm = km.convert(LengthUnit::Centimeter);
-        assert_eq!(1500.0, m.value());
-        assert_eq!(150000.0, cm.value());
-    }
-}
